requirements.txt
django==3.1
django-cors-headers==3.6.0
djangorestframework==3.12.4
PyYAML==5.3.1
gunicorn==20.0.4
djangorestframework-simplejwt==4.7.1
